Fix matrix-multiply breakage on 32-bit

+++ base/linalg_dense.jl
@@ -138,7 +138,7 @@ function _jl_generic_matmatmul{T,S,R}(C::StridedMatrix{R}, tA, tB, A::StridedMat
     mB, nB = lapack_size(tB, B)
     if nA != mB; error("*: argument shapes do not match"); end
     if size(C,1) != mA || size(C,2) != nB; error("*: output size is incorrect"); end
-    tile_size = ifloor(sqrt(tilebufsize/sizeof(R)))
+    tile_size = int(ifloor(sqrt(tilebufsize/sizeof(R))))
     sz = (tile_size, tile_size)
     Atile = pointer_to_array(convert(Ptr{R}, pointer(Abuf)), sz)
     Btile = pointer_to_array(convert(Ptr{R}, pointer(Bbuf)), sz)
